Moved endpoint.go to avoid internal ref from /pkg
